refactor: extract unsaved change guard hook
